% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_fonctions.R
\name{calcul_stats}
\alias{calcul_stats}
\title{Calculer des statistiques groupées}
\usage{
calcul_stats(data, group_var, numeric_var)
}
\arguments{
\item{data}{Jeu de données}

\item{group_var}{Variable de regroupement (ex : "color")}

\item{numeric_var}{Variable numérique (ex : "price")}
}
\value{
Un tableau avec moyenne, médiane et maximum
}
\description{
Calculer des statistiques groupées
}
