% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamonds.R
\docType{data}
\name{diamonds}
\alias{diamonds}
\title{Jeu de données diamonds}
\format{
Un data frame avec X lignes et Y colonnes :
\describe{
  \item{carat}{Poids du diamant}
  \item{cut}{Qualité de coupe}
  \item{color}{Couleur}
  \item{clarity}{Pureté}
  \item{price}{Prix en dollars}
}
}
\source{
Données fournies dans le cadre du cours
}
\usage{
diamonds
}
\description{
Un jeu de données contenant des informations sur des diamants.
}
\keyword{datasets}
