% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_fonctions.R
\name{filtrer_diamonds}
\alias{filtrer_diamonds}
\title{Filtrer le jeu de données diamonds}
\usage{
filtrer_diamonds(data = diamonds, cut = NULL, color = NULL)
}
\arguments{
\item{data}{Jeu de données (par défaut : diamonds)}

\item{cut}{Valeur(s) de coupe à sélectionner (optionnel)}

\item{color}{Valeur(s) de couleur à sélectionner (optionnel)}
}
\value{
Un data frame filtré
}
\description{
Cette fonction filtre le jeu de données selon la couleur et/ou la qualité de coupe.
}
