% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_fonctions.R
\name{plot_stats}
\alias{plot_stats}
\title{Graphique ggplot2 basé sur les statistiques groupées}
\usage{
plot_stats(stats_table, group_var)
}
\arguments{
\item{stats_table}{Tableau produit par calcul_stats()}

\item{group_var}{Nom de la variable de regroupement}
}
\value{
Un graphique ggplot2
}
\description{
Graphique ggplot2 basé sur les statistiques groupées
}
