% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_fonctions.R
\name{filtrer_diamonds}
\alias{filtrer_diamonds}
\title{Filtrer le jeu de données diamonds}
\usage{
filtrer_diamonds(data = diamonds, cut = NULL, color = NULL)
}
\arguments{
\item{data}{tibble diamonds}

\item{cut}{filtre coupe (NULL = tout)}

\item{color}{filtre couleur (NULL = tout)}
}
\value{
tibble filtré
}
\description{
Filtrer le jeu de données diamonds
}
